+++ src/discrete/Domain.kt
@@ -12,6 +12,7 @@ class Variables {
 
 class Domain {
 	companion object {
+		fun <T> from(items: List<T>): List<Value> = (0 until items.size).toList()
 		fun boolean(): List<Value> = listOf(1, 0)
 		fun range(first: Int, last: Int): List<Value> = (first..last).toList()
 		fun index(numItems: Int): List<Value> = (0 until numItems).toList()