+++ src/discrete/Solution.kt
@@ -14,4 +14,6 @@ class Solution {
 
 	val values: List<Value>
 		get() = map.values.toList()
+
+	fun tuple(p: Problem): List<Value> = p.variables.map { this.map[it] ?: 0 }
 }